Refactor: authentication page 🪚🔨🪄
